#22 Change Speed Controller Into Generic Runtime Robot Parameters
Also fix various styling issues.
