+++ src/scenes/Cluster/index.js
@@ -76,12 +76,14 @@ const simConfig = {
   env: {
     width: 600,
     height: 400,
-    speed: 15,
-    renderSkip: 1,
+    renderSkip: 1
   },
   robots: {
     count: 10,
     radius: 10,
+    params: {
+      velocityScale: 15
+    },
     controllers: {
       velocity: {
         controller,
@@ -119,10 +121,13 @@ const benchmarkConfig = {
       name: '5 Robots',
       simConfig: {
         env: {
-          speed: 50
+          renderSkip: 50
         },
         robots: {
-          count: 5
+          count: 5,
+          params: {
+            velocityScale: 50
+          }
         }
       }
     },
@@ -130,7 +135,12 @@ const benchmarkConfig = {
       name: '20 Robots',
       simConfig: {
         env: {
-          speed: 50
+          renderSkip: 50
+        },
+        robots: {
+          params: {
+            velocityScale: 50
+          }
         }
       }
     }
