+++ src/scenes/Demo/index.js
@@ -60,7 +60,6 @@ const simConfig = {
   env: {
     width: 600,
     height: 400,
-    speed: 15,
     renderSkip: 1,
     fields: {
       heatMap: {
@@ -72,6 +71,9 @@ const simConfig = {
   robots: {
     count: 3,
     radius: 15,
+    params: {
+      velocityScale: 15
+    },
     controllers: {
       velocity: {
         controller,
@@ -114,10 +116,13 @@ const benchmarkConfig = {
       name: '5 Robots',
       simConfig: {
         env: {
-          speed: 50
+          renderSkip: 50
         },
         robots: {
-          count: 5
+          count: 5,
+          params: {
+            velocityScale: 50
+          }
         }
       }
     },
@@ -125,7 +130,12 @@ const benchmarkConfig = {
       name: '20 Robots',
       simConfig: {
         env: {
-          speed: 50
+          renderSkip: 50
+        },
+        robots: {
+          params: {
+            velocityScale: 50
+          }
         }
       }
     }
